# Flush live and archive state in parallel