This should improve the flush performance on high-latency disks - flushing archive and live state in parallel.