@hkalina 2024-04-29
Benchmark from https://github.com/Fantom-foundation/Carmen/pull/842/files with `numAccounts` increased to `1_000_000` on a Cloud VM (Google e2-standard-4 + Balanced persistent disk):

BEFORE:
```
jk@jk-benchmark:~/Carmen/go$ go test ./database/mpt -run none -bench BenchmarkMptState_ArchiveFlush --benchtime 60s
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Carmen/go/database/mpt
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
BenchmarkMptState_ArchiveFlush-4              	     420	 162800937 ns/op
BenchmarkMptState_ArchiveFlushForcedDirty-4   	      19	3350841151 ns/op
PASS
ok  	github.com/Fantom-foundation/Carmen/go/database/mpt	428.182s

jk@jk-benchmark:~/Carmen/go$ go test ./database/mpt -run none -bench BenchmarkMptState_ArchiveFlush --benchtime 60s
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Carmen/go/database/mpt
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
BenchmarkMptState_ArchiveFlush-4              	     424	 159765255 ns/op
BenchmarkMptState_ArchiveFlushForcedDirty-4   	      19	3358839753 ns/op
PASS
ok  	github.com/Fantom-foundation/Carmen/go/database/mpt	427.623s
```

AFTER:
```
jk@jk-benchmark:~/Carmen/go$ go test ./database/mpt -run none -bench BenchmarkMptState_ArchiveFlush --benchtime 60s
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Carmen/go/database/mpt
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
BenchmarkMptState_ArchiveFlush-4              	     420	 160730048 ns/op
BenchmarkMptState_ArchiveFlushForcedDirty-4   	      20	3282603372 ns/op
PASS
ok  	github.com/Fantom-foundation/Carmen/go/database/mpt	428.036s

jk@jk-benchmark:~/Carmen/go$ go test ./database/mpt -run none -bench BenchmarkMptState_ArchiveFlush --benchtime 60s
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Carmen/go/database/mpt
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
BenchmarkMptState_ArchiveFlush-4              	     408	 161312931 ns/op
BenchmarkMptState_ArchiveFlushForcedDirty-4   	      19	3364868670 ns/op
PASS
ok  	github.com/Fantom-foundation/Carmen/go/database/mpt	429.339s
```

Seems there is no visible performance benefit, even after increasing the `numAccounts` - closing this PR.