@hkalina 2024-05-01
As noted by Kamil, the previous benchmark was wrong - not actually using the archive. I have re-done this benchmark and results are much better:

Google Cloud, e2-standard-8, Balanced persistent disk

BEFORE:
```
1000 updates:
go test ./state -run none -bench BenchmarkFlushGoState --benchtime 10s
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Carmen/go/state
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
BenchmarkFlushGoState-8   	     100	 427674303 ns/op
PASS
ok  	github.com/Fantom-foundation/Carmen/go/state	50.488s

10_000 updates:
jk@jk-benchmark:~/Carmen/go$ go test ./state -run none -bench BenchmarkFlushGoState --benchtime 10s
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Carmen/go/state
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
BenchmarkFlushGoState-8   	     100	5048760126 ns/op
PASS
ok  	github.com/Fantom-foundation/Carmen/go/state	570.991s
```
AFTER:
```
1000 updates:
jk@jk-benchmark:~/Carmen/go$ go test ./state -run none -bench BenchmarkFlushGoState --benchtime 10s
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Carmen/go/state
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
BenchmarkFlushGoState-8   	     100	 296230634 ns/op
PASS
ok  	github.com/Fantom-foundation/Carmen/go/state	37.063s

10_000 updates:
jk@jk-benchmark:~/Carmen/go$ go test ./state -run none -bench BenchmarkFlushGoState --benchtime 10s
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Carmen/go/state
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
BenchmarkFlushGoState-8   	     100	3237959108 ns/op
PASS
ok  	github.com/Fantom-foundation/Carmen/go/state	387.023s
```
**This change reduces the flush time on cloud deployment by 30-36%.**